refactor: 룰과 보드 로직 분리
